Remove nonsense condition for console shortcuts
Supposedly console keyboard shortcuts (e.g. Ctrl+L to clear) could be
processed even if the console is not open, while not connected to a
server. Of course they should not work if the console is not open. I
believe this can never happen for our cgame because KEYCATCH_UI should
always be active in the main menu.