Tweak alt-tab and shift-escape key up behavior
Previously an alt-tab window focus changed could do be performed on a
key up event by holding down tab, then holding
down alt, then releasing tab. That behavior is inconsistent with other
programs (at least on Windows), and no other shortcut in Daemon works on
key up. So do it on key down only.

Also don't suppress the escape key up event when shift is held. It is
possible that the user pressed Escape before Shift, meaning that a
normal key down event for Escape was generated. So generate a key up
event for Escape in case that happened.