Tweak and refactor key repeat handling
Previously, additional key down events would be processed from pressing
and holding a key iff a menu or console is open. Instead, always
generate an event, but disable certain actions (including binds, as
befroe) if it's a repeat. The behavior is slightly different and
hopefully more consistent. Before, a repeat event could cause the engine
to decide that a key is down, but only if the console/menu is open. This
can be observed, for example, in the trap_Key_KeysDown cgame api. Now a
repeat event never causes the key state to be changed to down. I don't
know of any user-visible consequences.