Allow keyboard input to UI and binds simultaneously
The cgame can now, if it wishes, receive key up and key down events
while binds are active. Upon receiving a key down event, the cgame will
return a boolean indicating whether the keystroke was consumed by the
cgame. If the keystroke was consumed by the cgame, no binds will be
executed.

Requires breaking ABI changes to the key event and key catcher
interfaces.

Also I added info on whether a key down is repeating to the cgame event.
No plans to use it now but it might be useful someday.

It is now the cgame's responsibility to open the menu when Escape is
pressed. (Previously it could not detect that keystroke.)