renderer: implement high precision rendering framebuffers
This is needed to avoid color-banding with the various
division/multiplication round trips of the complete overBright
implementation. This will also be needed for sRGB colors in the
future too. It is enabled by default if a feature requires it.

This is optional and can be disabled with r_highPrecisionRendering,
or automatically disabled if the hardware doesn't support the feature.