introduce plane_t struct with normal and dist members
Previously the plane was stored in a vec4_t,
with the three first components being the normal,
and the last one being the distance.

A lot of code was running the vec3_t specific
macros on the vec4_t, it only worked because
we use unsafe types.

This change is a step in the road of making safe
types like vec2_t, vec3_t, vec4_t and turn macros
like VectorCopy into type-safe functions.

This also makes the code more explicit.

Instead of doing:

    vec3_t normal = { 0, 0, 0 };
    vec4_t plane;
    VectorCopy( normal, plane );

We can now do:

    vec3_t in = { 0, 0, 0 };
    plane_t plane;
    VectorCopy( normal, plane.normal );

The day VectorCopy is a type-safe function
instead of a macro, it will work.

There already exist a cplane_t struct
that carries normal and dist but also
more information, we may modify it to
use plane_t for normal and dist, this
would allow us to do:

    PlaneCopy( cplane.plane, plane )

instead of:

    VectorCopy( cplane.normal, plane.normal );
    plane.dist = cplane.dist;

But for now this is out of topic.