Introduce /injectFault; replace "developer" cvar
There was a cvar named "developer" which enabled 3 fault testing
commands: /error, /crash, and /freeze. This cvar was annoying because it
only worked if set on the command line.

Remove that cvar and the old fault injection commands and replace it
with the injectFault/cgame.injectFault/sgame.injectFault commands. This
allows testing the previously supported kinds of errors plus several
others. There is the possibility to inject faults in the gamelogic now.
There is an option to make the fault happen in a new thread.

The "developer" cvar seemingly had the purpose to prevent it from being
too easy to run a crash command by accident. With /injectFault it is
also made sufficiently difficult by requiring you to correctly spell the
desired type of error as the first argument.