Rework CVAR_INIT; migrate cvars to new-style
Make a new implementation (Cvar::INIT) to replace CVAR_INIT, which will
work for new-style cvars. CVAR_INIT worked by forbidding changes after
the cvar is registered with Cvar_Get(). Cvar::INIT allows the cvar to be
set on the command line, but nowhere else. For cases where you want the
cvar to continue to be settable until a bit later (might be useful for
setting it from a config file), the INIT flag can be added later at the
moment the cvar is first used.
The 2 cvars using CVAR_INIT have been migrated to new-style.