Remove useless attempt at alias cycle detection
There was some code attempting to detect when an alias command calls
itself. It cannot work because it is written under the assumption that
commands invoved by other commands are resolved recursively (like the
C++ stack depth would be proportional to the level of recursive
commands), but in fact such commands are prepended to a queue and
resolved iteratively. In this iterative paradigm detecting recursion
would be more work and is not worth doing.