Fix sript args with 'exec' command
The exec command is supposed to allow args for the script like exec
script.cfg arg0 arg1 arg2... which are set in cvars. It wasn't working
because the args were set after executing the script, due to the set
commands being 2nd in the command buffer.

This could be fixed by putting in the set commands last (making them
first in the buffer), but they don't really need to be buffered anyway
so I used Cvar::SetValue directly.