Add -connect variant that needn't be the last arg
Intended to be used with a protocol handler that routes through the
updater. If the updater gets the "URL" first, it can validate it so we
don't need to ignore further arguments as a defense against bad input.

This improves on -connect in case the user uses the command line
customization feature of the updater like %command% -foo +bar. The
-connect-trusted arg could be inserted in %command% without cutting off the
rest.

I'm intentionally leaving the original variant in; it may be useful to have
both possibilities.