Use std::set_new_handler to send OOM message
Instead of defining custom new and delete to ensure that an allocation
failure results in a fatal error with the "Out of memory", use
std::set_new_handler to accomplish the same. This std lib facility
allows setting a custom behavior (instead of throwing std::bad_alloc) on
failure of operator new.

This is preferable since we don't have to worry about proliferation of
`new` variants or messing up the sanitizers.