Avoid redundant zeroing in some renderer allocations
ri.Z_Malloc ultimately called through calloc and so received zeroed
memory. But the zeroing is redundant in many cases because the code
immediately either memsets it, or copies something into it. Introduce
Z_AllocUninit and Z_Calloc to begin to categorize the cases as to
whether they need zeroing.

Note: the Z prefix makes no sense as we have no "zone" allocator, but I
continue to use it as I don't want to have to search out the
corresponding free function for each allocation to rename it to
correspond to a renamed allocation function.