Loosen trace test tolerance + test planes
With SSE disabled, the default tolerance of EXPECT_FLOAT_EQ was too
tight for some tests to pass. (The non-SSE results are probably actually
closer to the true value; the difference probably arises from using
80-bit-precision x87 operations). The expected value was set by just
running the code and seeing what it output. Independent of compiler or
OS, the x87-based results appear to form a separate cluster from the SSE
ones.

Also test the plane returned by the trace in a couple tests.