Move MSVC warning disables from Compiler.h to cmake
Like GCC warnings, warnings will now be controlled by the build
system instead of disabled by including a file. This lets warnings for
Daemon and Unvanquished be controlled more independently.

Some of the warning disable flags will need to be duplicated in
Unvanquished's CMakeLists.txt since the build was relying on Compiler.h
to disable them.