Convert most vector macros into functions
This is good because it prevents multiple evaluation of the arguments.
As a live example of this bug, the CG_CalculateTimeFrac inside a
VectorMA in CG_RenderParticle in Unvanquished/src/cgame/cg_particles.cpp
would be evaluated 3 times.

Also it may be useful to copy a glm::vec3 returned from a function into
a vec3_t without needing a temporary variable.