Get rid of global tess.vboVertexSprite
The global variable tess.vboVertexSprite was only set by shaders
interested in using it and never cleared by Tess_Begin or anything.
GLShader::SetRequiredVertexPointers always checked for it and would
request ATTR_ORIENTATION if true. So sometimes this attribute was
requested by mistake for shaders that don't use it and vertex buffers
that don't provide it.

Use local variables instead and explicitly pass in a boolean to
SetRequiredVertexPointers when needed.