Fix buggy config string transmitting ('bcsN')
The bcs0, bcs1, and bcs2 commands are used to transmit an info string
that has more than a certain length. The old code chopped them into
pieces based on the original length rather than the escaped length and
provided only a fixed amount of overhead for escape characters. Turns
out the overhead wasn't enough for real-life use cases (there are a lot
of backslashes due to infostring separators) and so the command to send
the serverinfo configstring would sometimes overflow and be silently
dropped.

Fix this by chopping up the string based on the escaped length.