Rename variables vboVertexSprite -> vertexSprite
Remove "vbo" from the variables named vboVertexSprite, which tell if a
shader uses a deform vertexes autosprite or autosprite2 mode.
refEntity_t's of type RT_SPRITE can use this with vertexes specified by
the cgame, not from a static VBO. Cf. 'vboVertexSkinning' which is
actually false when vertex animation is done without VBOs.