Tessellation life cycle cleanup
In Tess_Begin, warn if there is leftover data that was not consumed.
Clean up some cases where this can happen by ensuring that Tess_Clear or
Tess_End is always called after doing something.

Add Tess_StageIteratorDummy which does nothing except issue a warning
when it is called. Used for cases where overflow is not supposed to
happen and would not be handled correctly.