Material system: fix missing surfaces from command buffer
Fixes part (1) of #1259 (some surfaces can be randomly missing,
fixed at load time).
The problem was it did not ensure a unique value of tr.viewCountNoReset
which is used to mark surfaces as having been visited. The visit counter
is also used by R_MarkFragments which can be called before the 1st frame
is rendered.