tr_bsp: do not load deluxe maps if deluxe mapping was disabled because of hardware being incompatible
Fixup for #1229:

- https://github.com/DaemonEngine/Daemon/pull/1229

The engine now knows how to disable deluxe mapping when the r_deluxeMapping cvar
is enabled while the hardware is known to not be able to support the feature.

The renderer always loaded the deluxe maps when that cvar was enabled and told
the renderer the feature was available, meaning it would select the deluxe mapping
shader while it has never been built if the hardware doesn't support the feature.