Nuke R_AddLightBoundsToVisBounds()
visBounds is used only to choose zFar. A comment says that
expanding the vis bounds was needed for occlusion queries (which
were NUKED in 2014). I looked at the nuking commit
ffd171413eab21bf81a1d8c7f456eef90a243bc3 and it seems that the occlusion
code involved GL draw calls to "draw" the boundaries of the light and
check whether anything would show up. So it's plausible that the
expanded zFar would be needed for that stuff. But it shouldn't be needed
now.