NUKE static lights (realtime lights from BSP ents)
A "static light" was a type of realtime light. It was instantiated by
a BSP entity with classname "light", which must have
the "noradiosity" keyword set in order to be rendered. Judging by
https://users.unvanquished.net/~slipher/map-entity-directory.txt there
are no Tremulous or Unvanquished maps using this.

90% of the code for this feature was optimizations specific to the
deprecated forward renderer. (Optimizations in the sense that the code
is intended to be faster than the dynamic light code path.) If we ever
wanted to test real-time lights from map entities later, it could be
done by just adding a dynamic light based on the BSP entity. The entity
parser would be the only part that's needed.

Closes #1051.