Rename r_dynamicLight -> r_drawDynamicLights, unlatch
Giving the cvar for whether to add dynamic lights at runtime a different
name from r_realtimeLighting's old name will help it to avoid getting
mistakenly set by old autogen.cfg values. The name name follows the
convention that r_drawXXX cvars that can be used to disable something
that is usually rendered. (showXXX cvars on the other hand are cheat
cvars for drawing a representation of something that is usually unseen.)

Remove Cvar::Latch so it can be toggled instantly.