Default heat haze shader sort to BLEND0
Make the automatically determined sort order for shaders containing a
HEATHAZEMAP BLEND0 (equal to typical translucent surfaces).

Some of our Unvanquished assets, such as the lcannon impact shockwave
shader, were usually not rendered due to a bad sort value (SS_OPAQUE).
Only the flamer ones currently appear to work correctly. This is because
the flamer heathaze shaders are registered after the renderer loads the
map (unlike the lcannon ones). Within the SS_OPAQUE group, surfaces are
sorted in the order their shaders are registered. So the flamer ones
mostly worked because they could at least distort opaque map surfaces
correctly. Meanwhile the lcannon ones were so early in the sort order
that the sky was the only thing they could manage to distort.