Drop f16_t from shaderVertex_t; half float now optional
shaderVertex_t is the type used for vertex data which is generated and
uploaded to the GPU on every frame. The motivation behind getting rid of
the half floats is to run on graphics backends that don't support half
float vertex data. I assume there is not much benefit to using half floats
here: normally you would use them to save memory in data which is stored
long-term. But with shaderVertex_t we're only uploading a few thousand
verts each frame and then throwing them away. Though systems that aren't
able to run all skeletal models on the GPU could be an exception as they
have to additionally upload a lot of model verts each frame.

GL_ARB_half_float_vertex is now an optional extension.