Revert bdb47eb9 computeLight_fp refactoring
Revert that commit and the two subsequent ones which conflict.

It changed how the dynamic lights are distributed into layers: before,
the layer was set by the light's index modulo 4. After, the first
quarter of lights are in layer 0, the 2nd quarter in layer 1, etc. The
distribution is (hopefully) 'random' with the previous algorithm. With
the new algorithm they are grouped by when they are added. So the new
algorithm is more likely to fail to render some lights because all the
lights in an area ended up in one layer instead of being well
distributed. I have prepared a trail system demonstrating this which I
will attach to the PR.

Also the commit makes the GLSL fail to compile on my machine. I had to
add a float cast to the following line:
	const vec3 tileScale = vec3( r_tileStep, 1.0 / numLayers );

------

Revert "Cleanup nuke non-UBO and non-integer texture paths"

This reverts commit f7e67a10e21a82024ba590f3b0de346e652ba917.

Revert "Tabs + spaces -> tabs in lighttile_fp"

This reverts commit 7d793e95e896fb1989ebe293b0edc7e283c10b16.

Revert "Optimise and cleanup lighttile"

This reverts commit bdb47eb95710d797d9f5267b87d940b88bf80962.