Implement static reflection cubemap caching
Add `r_autoBuildCubeMaps`.

0: off
1: cached on disk
2: always rebuild when loading the map

This saves the grid itself, cubemap probe list with positions, and cubemap images in ktx format. The grid file also contains version, some BSP header lump lengths (to minimise the possibility of getting wrong cubemaps due to a map update), as well as cubemap size and spacing. If any of these don't match, the cubemaps have to be rebuilt.