Fix heatHaze with material system
Do a second rendering for heatHaze materials. This is done instead of making heatHaze into to stages to avoid duplicating the memory (+possible extra material padding).

Added a `u_DeformEnable` uniform to the material heaHaze shader variant, which is global, because `u_DeformMagnitude` is set to 0.0 for the second rendering of the surface/surfaces.