Split server.private in 2 cvars; fix LAN game not listable
illwieckz discovered that the default value 3 of cvar server.private
on clients made it impossible to see the game in the LAN server
list. This mode 3 really makes no sense as it is allowing LAN clients
only, but not even allowing them to see the server.

server.private was mixing two things: whether the server reports
its status and whether it allows non-LAN addresses. Make it control only
the former and add sv_networkScope for whether non-LAN addresses are
allowed.