Clean up snapshot mess with #including gamelogic
A shared header defined snapshot_t which was used in both the client
and cgame, but with different definitions of the player state struct.
For the cgame definition of playerState_t it included the file
shared/bg_public.h from the gamelogic.
Instead of this define an ipcSnapshot_t type which has a consistent
definition and is used for IPC messages. Now snapshot_t will be defined
and used only in the cgame.