Add unit tests for most q_math transform_t functions
And fix a bug in one implementation of TransInsScale (this function is
not used in the codebase).

These transform_t functions deal with composing rotation, translation,
and scale operations. So transform_t represents a similarity-preserving
transformation. All of the q_math transform functions get a test, except
the 3 that have to do with interpolation (I'm not sure how that is
supposed to work).