Rework SSE intrinsics to avoid Clang 19 bugs
Fixes the Clang part of
https://github.com/DaemonEngine/Daemon/issues/1312. The problem is that
LLVM wrongly considers __m128 to have floating-point semantics at all
times. When it sees an all-ones mask in some component it wrongly sees
it as a NaN and replaces it with an undefined value when in
-ffinite-math mode.

In the cases where the new first_XYZ_second_W is used, that is actually
what Clang itself outputs after optimizing the original intrinsics code
(with a non-bugged flag configuration). So hopefully it is faster
anyway.