Fix daemonded not starting on Mac (socket path too long)
The default profile now generates a ~50-character-long $TMPDIR by
default (I assume it must have been shorter before). The singleton
socket path we generate for daemonded ends up too long for the
`bind` or `connect` APIs. (The socket path for clients is slightly
shorter and just sneaks under the limit.) Work around this by using
the Mac-specific pthread_chdir_np API to thread-safely change the
CWD to the directory with the socket, accessing it with a relative
path, then resetting the current directory to what it was before.

Fixes #1390.