Define appropriate flags for all engine commands
Commands can have flags such as 'BASE' or 'RENDERER' which determine the
command listing commands that they appear in. If a command has no flags
at all, it does not appear in /listCmds, which seems bad. So I disabled
the two-argument StaticCmd constructor without flags in the engine, and
defined flags for all the commands using it.

Also I changed the set of available flags a bit. Nuked SYSTEM because
it is too vague, and added SERVER and CLIENT.