Remove map_restart + sv_maxclients change special case
The special case was that if map_restart is used after sv_maxclients is
modified, the server does a full map change (i.e. clients have to
reload) instead of the normal map_restart. I don't think this is a great
idea since the map_restart semantics may be important for example if it
is used to implement warmup. Plus it makes implementing the cvar more
complicated.