Fix moronic light grid array size calculation
First it calculated the number of points along the negative
and positive axes of a dimension of the light grid. Then, instead
of just adding those two together to determine the total number of
points on the axis, it calculated the coordinates of the minimum and
maximum points, subtracted them, and divided by the cell size to
get the number of points on the axis. This division was floored, not
rounded, so if the difference in coordinates had an error of -0.000001
it would get the wrong answer.

Surprising that this didn't break earlier.
Fixes #1518.