Fix CGEN_IDENTITY_LIGHTING with 2D drawing
Some 2D shaders are automatically assigned CGEN_IDENTITY_LIGHTING. This
causes a bug of the cursor being black in the main menu (where
tr.identityLighting is 0) and the wrong color if r_overbrightQ3 is used
while in a map. Fix by making it behave as CGEN_IDENTITY during 2D
rendering.

Fixes #1590.