Use a simple screenSpace vertex shader
Changes `GLShader` constructors, removing unnecessary stuff from them, and making it only have 2 distinct constructors - for rendering shaders and compute shaders.

Changes the relevant shaders to use a new `screenSpace` shader, which is almost a no-op.

NUKE the the other .glsl shaders that just repeat the same functionality.

Also NUKES the `u_ModelViewProjectionMatrix` from them, because the result is always a full-screen quad anyway.

Adds `Tess_InstantScreenSpaceQuad()` for use with this, because with some other shaders we still do need `u_ModelViewProjectionMatrix`.