DaemonArchitecture: NUKE x86 revision detection
Always assume i686, instead of trying to use the numeric value of
Q_PROCESSOR_X86, which isn't accurate in our build.

For architecture detection the compiler is invoked without any of the
flags that we usually pass. So the -march=XXX, etc. flags are not there.
This means the architecture revision that the compiler advertises during
architecture detection may not be the same as in the real build. You may
react, "just do the architecture detection with all flags on", but this
would probably lead to chicken-and-egg problems as architecture is used
to choose flags. In principle it could be done with two passes of
architecture detection but implementing that would obviously be a waste
of time.

Fixes #1596.