Add cvar to disable native exe crash handling
So server owners who want to get Unix core dumps for native exe
gamelogic could set sgame.nativeExeCrashHandler to 0. Doing this stops
the crash handler from being installed. In case of a fatal C++ exception
we still install the terminate handler so we can log the exception
message, but afterward, if the cvar is off, we call abort() instead of
shutting down with the ErrorMsg IPC.