external_deps: don't munge SDL config
The build script for mingw sdl2 uses prebuilt binaries, but calls a
makefile which only substitutes some hard-coded paths into some build
config helpers and copies the files. Just copy the files ourselves and
don't munge the configs: the unmunged cmake uses a relative path
which is actually what we want. In newer SDL versions the makefile munges the
cmake configs (in addition to the shell-based one) which breaks us.