Improve lazy shader building
Previously, when `r_lazyShaders 1` was used, it would simply load all of the possible shaders with the current graphics settings, and without deform shaders.

With this change, it will instead mark shaders that are used for BSP and model surfaces for building, which will be done in `RE_EndRegistration()`. Skins are also loaded (the variants with skeletal animation are always loaded because we don't know if an MD5 model will need them later).

This reduces the amount of shaders built, and allows building deform shaders at map load.