Don't search generated textures in shader hash table
RE_RegisterShaderFromImage creates a q3shader object for textures that
have just been generated at runtime. Don't attempt to find
them in the shader hash table. This doesn't make much sense and even if
a match were found, an image was already uploaded at that point so the
memory is already wasted.