Allow shaders to be re-registered with different flags
Allow instantiating multiple versions of a q3shader if it was registered
with different RSF_* flags each time. This will make it so that
registering stuff in a different order will not give different behavior.

NUKE shaderType_t and replace it with a new flag RSF_3D that
replicates the behavior of the old shaderType_t::3D. Previously, shaders
could be registered multiple times but only based on a differing
shaderType_t.