Fix SSAO + material system on AMD
Fixes #1616. Same trick as #1704 for fixing depth fade - insert a
texture barrier so that reading the depth texture is valid.

Some depth-writing stuff such as polygonOffset decals and alpha-tested
textures is rendered after SSAO is done, so in general we need to do the
texture barrier twice (if SSAO is on).