sdl3: disable the "accent menu" on macOS
For some reason SDL3 made a change to enable a system menu offering
to input a letter with diacritics, which appears whenever you press
and hold a key (for example while running). After SDL does this,
call the same Objective C facility to undo the configuration change
that SDL did.