sdl3: fix r_displayIndex on Linux
First of all, SDL always centers the window on all our platforms
regardless of whether "centered" or "undefined" position is requested.
It turns out that on Linux, requesting centering on display N makes it
center on display N, while requesting any position on display N means
center on unspecified display. So NUKE r_centerWindow and always request
centering since not centering is never possible and this makes
r_displayIndex be honored on Linux.